Leather-defect/requirements.txt
streamlit>=1.25.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
dotenv
matplotlib>=3.7.0
scikit-learn>=1.2.0 
seaborn>=0.12.0